requirements.txt
click==6.3
gitdb==0.6.4
GitPython==1.0.2
smmap==0.9.0
wheel==0.24.0